+++ src/main/java/com/example/WBMdemo/entity/Material.java
@@ -29,12 +29,12 @@ public class Material {
 	@SequenceGenerator(name="material-seq-gen",sequenceName="MATERIAL_SEQ_GEN", initialValue=1, allocationSize=1)
 	@GeneratedValue(strategy = GenerationType.AUTO, generator="material-seq-gen")
 	@Column(name = "ID")
-	private Integer materialId;
+	private Long materialId;
 	
 	@Column(name = "MATERIAL_NAME")
 	private String materialName;
 	
 	@Column(name = "MATERIAL_PRICE")
-	private Long materialPrice;
+	private BigDecimal materialPrice;
 	
 }
